modal pops up with text, highlighting needs fixed, submit needs fixed
